メインゲームのスコアをフォーマット

+++ Assets/Scripts/Tools/ScoreFormatter.cs
@@ -5,7 +5,7 @@ namespace G2048.Tools
 {
     public class ScoreFormatter
     {
-        public static string Format(int score)
+        public static string Format(ulong score)
         {
             return string.Format("{0:N0}", score);
         }
